% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_multi_exp}
\alias{boxplot_multi_exp}
\title{boxplot of multiple species and different experiments}
\usage{
boxplot_multi_exp(list.spe, path, xlab, ylab, list.prior.part = NULL,
  logtrans = FALSE, rank)
}
\arguments{
\item{list.spe}{list. data.list}

\item{path}{string. result location}

\item{xlab}{string. x axis}

\item{ylab}{string. y axis}

\item{list.prior.part}{list. the prior range}

\item{logtrans}{bool. whether use log transform}

\item{rank}{array. rank of presenting the data}
}
\value{
just plot no return
}
\description{
can be parameters or intial value
prior range of species can be also presented
}
\seealso{
[boxplot_multi_exp_comb()]
}
