% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change_block}
\alias{change_block}
\title{Editing or showing specific blocks (ens.i01)}
\usage{
change_block(paternlist = NULL, contentlist = NULL, infile = NULL,
  outfile = NULL, type = "show")
}
\arguments{
\item{paternlist}{list. indicates the location for block changing
->pattern0: the start of the block to be matched
->pattern1: the end of the block to be matched.
->shift: a vector. if length(shift)==2. Index shift will be added on pattern0 and pattern1
         if length(shift)==1. The shift is the same for both pattern0 and pattern1
must be provided.}

\item{contentlist}{list. the value to be changed
->file: the file to read from
->ilinerag: the index range of lines as input
Default NULL}

\item{infile}{string. input file path. must be provided}

\item{outfile}{string. output file path. default NULL}

\item{type}{string. "show"(show the parameter, default), "edit"(edit the value)
if there is mulitple value for the corresponding term, only show value is possible}
}
\value{
If type="edit", no value is returned and just change the file correspondingly
        If type="show", return the show part
}
\description{
The function will edit or show the value of specific blocks
}
\seealso{
[change_para()] for modifying one parameter
}
