% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{densi_region}
\alias{densi_region}
\title{constructed density frunction from a list of interval}
\usage{
densi_region(list.intev, logtrans)
}
\arguments{
\item{list.intev}{list. list of parameter range}

\item{logtrans}{bool. whether log transform the interval. TRUE log transformed}
}
\value{
list. list of range distribution information
}
\description{
constructed density frunction from a list of interval
}
