% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultformat.R
\name{depend_for}
\alias{depend_for}
\title{dependences formulation}
\usage{
depend_for(y, x, func)
}
\arguments{
\item{y}{array. the location of the dependent variable. must be provided}

\item{x}{array. the independent variable. must be provided}

\item{func}{array. the function to be formulated for the dependency. must be provided}
}
\value{
string. output information
}
\description{
automatic dependence formulation with default format. depricated soon.
}
\seealso{
[spec_output()] for producing species block in i01 for default format
         [react_output()] for producing reaction block in i01 for default format
}
