% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.hist.R
\name{draw_hist}
\alias{draw_hist}
\title{plot distribution histogram for summary}
\usage{
draw_hist(tab, loci, xlab)
}
\arguments{
\item{tab}{array. the data table}

\item{loci}{string. location for result(add on string for name)}

\item{xlab}{string. x axis}
}
\value{
no return just plot the histogram
}
\description{
plot distribution histogram for summary
}
