% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.sweep.R
\name{draw_sweep}
\alias{draw_sweep}
\title{plot var~sweep}
\usage{
draw_sweep(tab, para, loci, linethick)
}
\arguments{
\item{tab}{array. the data table}

\item{para}{string. y axis}

\item{loci}{string. location for result (add on string for name)}

\item{linethick}{bool. whether make line thicker}
}
\value{
no return just plot
}
\description{
plot the variable through sweeps
}
