% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp.model.overlay.R
\name{exp_model_overlay}
\alias{exp_model_overlay}
\title{overlay experiment measurement with model predicted result in o01}
\usage{
exp_model_overlay(o01, species, addon, exp, dir.res, logtrans = FALSE)
}
\arguments{
\item{o01}{string. file path}

\item{species}{string. the species that need overlay}

\item{addon}{string. addon on output file name}

\item{exp}{numeric. the experiment number that need to be plot}

\item{dir.res}{string. the output folder}

\item{logtrans}{bool. whether log transformaed is performed on dat}
}
\value{
figure. return the figure itself
}
\description{
measurement is actually "scaled" measurement
the range [low high] will also be plot
}
