% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table_reader.R
\name{exp_model_tab}
\alias{exp_model_tab}
\title{experiment specie concentration fetching(o01)}
\usage{
exp_model_tab(lines = NULL, start = NULL, end = NULL,
  specname_ind = NULL, power_ind = NULL, ch_ind = NULL,
  species = NULL)
}
\arguments{
\item{lines}{array. lines to parse. must be provided}

\item{start}{array. start of the index range. must be provided}

\item{end}{array. end of the index range. must be provided}

\item{specname_ind}{array. the index of species part. must be provided}

\item{power_ind}{array. the index of power line. must be provided}

\item{ch_ind}{array. the index of the number line. must be provided}

\item{species}{string. the specie to fetch. must be provided}
}
\value{
array. the informaiton table
}
\description{
experiment specie concentration fetching from the table in file o01
}
\seealso{
[mod_model_tab()] for similar function on model estimation information
}
