% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{file_str_which}
\alias{file_str_which}
\title{file based str_which}
\usage{
file_str_which(path, pattern)
}
\arguments{
\item{path}{string. input file for the string}

\item{pattern}{string. pattern to be matched}
}
\value{
array. the index(line number) of matched position
}
\description{
file verion of str_which, used especiall for large files
}
