% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{foldcreate}
\alias{foldcreate}
\title{folder creating}
\usage{
foldcreate(dirpath = NULL)
}
\arguments{
\item{dirpath}{string. path to the folder to create. must be provided}
}
\value{
no return just create the folder
}
\description{
this function will create the folder if the folder doesn't exist
}
