% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formu_reac.R
\name{formu_reac}
\alias{formu_reac}
\title{formualte transcription and translation}
\usage{
formu_reac(enz = NULL, path = NULL)
}
\arguments{
\item{enz}{array. the enzyme list. must be provided}

\item{path}{string. the result location. must be provided}
}
\value{
no return
}
\description{
this is funciton is for producing transcription and translation reaction from a gene/enzyme list
}
