% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoge.file.R
\name{homoge_file}
\alias{homoge_file}
\title{modify output script for consistency}
\usage{
homoge_file(localpath, foldname)
}
\arguments{
\item{localpath}{string. the local working directory}

\item{foldname}{string. the name of the project folder}
}
\value{
print out result no return
}
\description{
makesure the output analysis script agreed with the original input file
}
