% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniguess.R
\name{iniguess}
\alias{iniguess}
\title{set initial guess}
\usage{
iniguess(input, output, valvec)
}
\arguments{
\item{input}{string. input file path}

\item{output}{string. output file path}

\item{valvec}{array. named vector of values for initial guess. block of value that wasn't inlcuded in the vector will be initilized as lowval + smallvalue}
}
\value{
no return just change the file
}
\description{
this function will use the vector to initilize the inital guess for parameters
}
