% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kine_para_refine.R
\name{kine_para_refine}
\alias{kine_para_refine}
\title{refine kinetic parameters}
\usage{
kine_para_refine(list.res.refine = NULL, range.speci = NULL,
  extendrag = NULL)
}
\arguments{
\item{list.res.refine}{list. the species non-specific data struct. must be provided}

\item{range.speci}{list. the species specific data struct. must be provided}

\item{extendrag}{numeric. extent to extend unknown range in kinetic parameters(when there is only one value in in record). must be provided}
}
\value{
list. refined parameter list
}
\description{
this function will take kinetic parameters structure and modify it
will calcualte range for all kinetic parameters and give default value for those without enough information
}
