% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{linecount}
\alias{linecount}
\title{count number of lines in the file}
\usage{
linecount(path)
}
\arguments{
\item{path}{string. the file location}
}
\value{
int. line number
}
\description{
this function will count number of lines in the file
}
