% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{multi_line_map}
\alias{multi_line_map}
\title{str_which for multiple lines}
\usage{
multi_line_map(string, pattern)
}
\arguments{
\item{string}{string. input string to match with. array with each element for a line}

\item{pattern}{string. the pattern to find. \\n is in the pattern}
}
\value{
array. the index location for the last line of the matching block
}
\description{
this function performs as str_which (no \n in pattern).
If there is \n in pattern, the function will try to match corresponding continuous line and return index of the last line matched.
}
\examples{
string=c("aabb","cc","eeab","ce")
pattern="b\\\\nc"
multi_line_map(string=string,pattern=pattern)
pattern="eea"
multi_line_map(string=string,pattern=pattern)
pattern="eea\\\\n"
multi_line_map(string=string,pattern=pattern)
}
