% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{nameparser}
\alias{nameparser}
\title{parser for name information}
\usage{
nameparser(lines, ind, pattern)
}
\arguments{
\item{lines}{array. specific lines to parse}

\item{ind}{array. the index to define start searching point}

\item{pattern}{string. the pattern for titile. 'namespec' or 'namereac'}
}
\value{
array. parameter name list
}
\description{
this function parse the part wtih species and reaction information
it will return names of theta.
only for o02 file 'namespec' and 'namereac' block
}
