% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filecleaner.R
\name{o02_clean}
\alias{o02_clean}
\title{clean o02 file}
\usage{
o02_clean(path)
}
\arguments{
\item{path}{string. input file}
}
\value{
string. the path to the cleaned file
}
\description{
this script is before fetching information out of output file from ensemble
currently only deal with RESTART run which cause some block rearrangement in o02 file
}
