% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filereader.R
\name{o02_reader}
\alias{o02_reader}
\title{reader for o02 file}
\usage{
o02_reader(path, flagsweep = FALSE)
}
\arguments{
\item{path}{string. input file}

\item{flagsweep}{bool. TRUE or FALSE.
TRUE: only the end of sweep part are considered for final reasult.
FALSE: output all blocks. Default FALSE}
}
\value{
list. list containing information in o02
}
\description{
file reader for o02 file
}
