% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filereader.R
\name{o03_reader}
\alias{o03_reader}
\title{reader for o03 file}
\usage{
o03_reader(pathlist = NULL, ioutselec = NULL)
}
\arguments{
\item{pathlist}{list. input file list for o03. must be provided}

\item{ioutselec}{int. the iout_id to be read. must be provided}
}
\value{
array. array of time series data. time x species x #blocks
}
\description{
file reader for o03 file
}
