% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi_spec_time.R
\name{plot_multi_spec_time}
\alias{plot_multi_spec_time}
\title{plot of time trajectory of multiple species}
\usage{
plot_multi_spec_time(o01 = NULL, species.list = NULL, addon = "",
  exp = NULL, dir.res = NULL, ncol = 5, logtrans = FALSE)
}
\arguments{
\item{o01}{string. the location of o01. must be provided}

\item{species.list}{array. the species to plot. must be provided}

\item{addon}{string. addon file name. default ""}

\item{exp}{int. the experiment number to choose. must be provided}

\item{dir.res}{string. result locaiton. must be provided}

\item{ncol}{int. the number of columne in plot_grid plot. default 5}

\item{logtrans}{bool. whether log transformation is  performed. default FALSE}
}
\value{
figure. the figure itself
}
\description{
plot of time trajectory of multiple species
}
