% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre.file.prepare.R
\name{pre_file_prepare}
\alias{pre_file_prepare}
\title{prepare input files}
\usage{
pre_file_prepare(dir.tempt, templatepath)
}
\arguments{
\item{dir.tempt}{string. the temptory directory}

\item{templatepath}{string. the template path}
}
\value{
just move the file no return value
}
\description{
thie function will copy all template file needed into the temptory directory
}
