% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filereader.R
\name{prior_reader}
\alias{prior_reader}
\title{read the prior information}
\usage{
prior_reader(path = NULL)
}
\arguments{
\item{path}{string. the location of i01 for prior information. must be provided}
}
\value{
list. list containing prioir information
}
\description{
read prior range of the enyzme initial concentration and kinetic parameters
initial value for enzyme is not correct if you are using dependence
there is some manual modification in i01 file, and so
there can be some differences in space number. so use '\s' to replace ' '
this function only works for the simple formulation of i01
for later version with much complex dependence formulation it is no longer working.
}
