% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randloglhs.R
\name{randloglhs}
\alias{randloglhs}
\title{initial guess generate}
\usage{
randloglhs(nsample, isample, input, output)
}
\arguments{
\item{nsample}{int. number of replicate to produce}

\item{isample}{int. the sample index}

\item{input}{string. input file path}

\item{output}{string. output file path}
}
\value{
just change the file no return
}
\description{
this function will generate the initial guess for all parameters by lhs (Latin Hypercube sample) in log-uniform distribution
the lhs matrix will be calcualted for isample and used afterwards
}
