% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.format.R
\name{react_output}
\alias{react_output}
\title{formulate the reaction part of i01 (in a default format)}
\usage{
react_output()
}
\value{
list. output information
}
\description{
formulate the species part for i01 file with default format
the argument is not set explicit and they are transferred by environment (the function that call react_output())
}
\seealso{
[react_output_format()] for similar function based on user defined format files
         [spec_output()] for producing species block in i01 for default format
}
