% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_format.R
\name{react_output_format}
\alias{react_output_format}
\title{formulate the reaction part of i01 based on format files}
\usage{
react_output_format()
}
\value{
list. input on file information and size
}
\description{
this function is indicated to depends on exsting format/template for each part
lines start with more than two '#' will be removed
template is needed
no input arguments as arguments are transferred by environment
}
\seealso{
[spec_output_format()] for producing species block in i01 for default format
         [react_output()] for similar function based on default format
}
