% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{read_reac}
\alias{read_reac}
\title{read reactions file}
\usage{
read_reac(path = NULL)
}
\arguments{
\item{path}{string. the path to the reaction list file. must be provided}
}
\value{
list. return the list containing c(reaction_name, substrates, products)
}
\description{
load the reaction list file
}
