% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{rela_ind_screen}
\alias{rela_ind_screen}
\title{use relative locaiton between indx and indy to filter indy}
\usage{
rela_ind_screen(indx, indy, rela = ">=")
}
\arguments{
\item{indx}{array. the reference index}

\item{indy}{array. the index to be filtered}

\item{rela}{string. the expected relative location between indx and indy}
}
\value{
array. the filtered indy
}
\description{
use relative locaiton between indx and indy to filter indy
}
