% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultformat.R
\name{spec_output}
\alias{spec_output}
\title{formulate the species part of i01(in a default format)}
\usage{
spec_output()
}
\value{
array. output information
}
\description{
formulate the species part for i01 file with default format
the argument is not set explicit and they are transferred by environment (the function that call spec_output())
}
\seealso{
[spec_output_format()] for similar function based on user defined format files
         [react_output()] for producing reaction block in i01 for default format
}
