% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_format.R
\name{spec_output_format}
\alias{spec_output_format}
\title{formulate the species part of i01 based on format files}
\usage{
spec_output_format()
}
\value{
array. input on file information
}
\description{
this function is indicated to depends on exsting format/template for each part
template is needed
no input arguments as arguments are transferred by environment
}
\seealso{
[react_output_format()] for producing reaction block in i01 for default format
         [spec_output()] for similar function based on default format
}
