% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_input}
\alias{summary_input}
\title{summary input files}
\usage{
summary_input(i01 = NULL, i02 = NULL)
}
\arguments{
\item{i01}{string. the locaiton of i01. must be provided}

\item{i02}{string. the locaiton of i02. must be provided}
}
\value{
list. list of information contains in input files
}
\description{
this funciton summary exist content in i01 and i02 file
output exist species and reaction
print imporant parameters
these information can be used for modify ens.def file
}
