% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_o02}
\alias{summary_o02}
\title{summary plot for o02}
\usage{
summary_o02(o02.data, dir.res, addonname, linethick = FALSE)
}
\arguments{
\item{o02.data}{list. the struct from o02.reader}

\item{dir.res}{string. the location for producing figures}

\item{addonname}{string. the addon name}

\item{linethick}{bool. whether thicker lines shoulde be draw for figure}
}
\value{
just plot no return
}
\description{
summary plot for o02
}
