% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{tabparse}
\alias{tabparse}
\title{table data parse}
\usage{
tabparse(lines_models = NULL, rowrange = NULL, colrange = NULL)
}
\arguments{
\item{lines_models}{array. the lines to fetch. must be provided}

\item{rowrange}{array. row range in the table. must be provided}

\item{colrange}{array. column range in the table. must be provided}
}
\value{
array. the selected part of the table
}
\description{
this funciton parse the table in input and output file and return specific column
space will be treated as separater
}
\examples{
lines_models=c("1 2 3 4 5","6 7 8  9 10","11 12 13\\t14 15")
tabparse(lines_models,c(1,3),c(2,4))
}
