% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{as_data_frame_rowname}
\alias{as_data_frame_rowname}
\title{modified version of as.data.frame for rownames}
\usage{
as_data_frame_rowname(list = NULL)
}
\arguments{
\item{list}{list. list to be converted. must be provided}
}
\value{
dataframe. will return t(as.data.frame(list))
}
\description{
modified version of as.data.frame to remove the changed caused by make.name
}
\examples{
inputlist=list("1.34"=c(1,2,3),"8.5"=c(3,2,3))
as_data_frame_rowname(inputlist)
}
