% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_multi_exp}
\alias{boxplot_multi_exp}
\title{boxplot of multiple species and different experiments}
\usage{
boxplot_multi_exp(list.spe = NULL, path = NULL, rank = NULL,
  xlab = "x", ylab = "y", list.prior.part = NULL, logtrans = FALSE)
}
\arguments{
\item{list.spe}{list. data.list. must be provided}

\item{path}{string. result location. must be provided}

\item{rank}{array. rank of presenting the data. must be provided}

\item{xlab}{string. x axis. default "x"}

\item{ylab}{string. y axis. default "y"}

\item{list.prior.part}{list. the prior range. default NULL}

\item{logtrans}{bool. whether use log transform. default FALSE}
}
\value{
just plot no return
}
\description{
can be parameters or intial value
prior range of species can be also presented
}
\seealso{
[boxplot_multi_exp_comb()]
}
