% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_multi_exp_comb}
\alias{boxplot_multi_exp_comb}
\title{boxplot of multiple species and different experiments}
\usage{
boxplot_multi_exp_comb(list.spe, path, xlab, ylab,
  list.prior.part = NULL, logtrans = FALSE, rank)
}
\arguments{
\item{list.spe}{list. list of data}

\item{path}{string. result location}

\item{xlab}{string. x axis}

\item{ylab}{string. y axis}

\item{list.prior.part}{list. the prior range}

\item{logtrans}{bool. whether use log transform}

\item{rank}{array. rank of presenting the data}
}
\value{
just plot no return
}
\description{
can be parameters or intial value
prior range of species can be also presented
the differences from boxplot.multi.exp is that this funciton will plot prior and result side by side
}
\seealso{
[boxplot_multi_exp()]
}
