% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{change_def}
\alias{change_def}
\title{editing or showing ens.def files}
\usage{
change_def(term, value, infile, outfile, type = "show")
}
\arguments{
\item{term}{string. the value that need to be changed}

\item{value}{array, dim(array)=1. the value}

\item{infile}{string. input file path}

\item{outfile}{string. output file path}

\item{type}{string. "show"(show the parameter), "edit"(edit the value)}
}
\value{
If type="edit", no value is returned and just change the file correspondingly
        If type="show", a list of value will be returned
}
\description{
the function is for editing or showing specific parameters
}
