% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combaccum.R
\name{comb_accum}
\alias{comb_accum}
\title{comb accumulated run}
\usage{
comb_accum(path = NULL)
}
\arguments{
\item{path}{string. the input o02 file. must be provided}
}
\value{
no return
}
\description{
this is for submitting multiple accumulation run individually and combine them by one run
the function will correct the index number and make the combined file looks like output of one run
}
