% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawhist.R
\name{draw_hist}
\alias{draw_hist}
\title{plot distribution histogram for summary}
\usage{
draw_hist(tab = NULL, loci = "", xlab = "xlab")
}
\arguments{
\item{tab}{array. the data table. must be provided}

\item{loci}{string. location for result(add on string for name). default ""}

\item{xlab}{string. x axis. default "xlab"}
}
\value{
no return just plot the histogram
}
\description{
plot distribution histogram for summary
}
