% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawsweep.R
\name{draw_sweep}
\alias{draw_sweep}
\title{plot var~sweep}
\usage{
draw_sweep(tab = NULL, ylab = "ylab", loci = "", linethick = FALSE)
}
\arguments{
\item{tab}{array. the data table. must be provided}

\item{loci}{string. location for result (add on string for name). default ""}

\item{linethick}{bool. whether make line thicker. default FALSE}

\item{para}{string. y axis. default "ylab"}
}
\value{
no return just plot
}
\description{
plot the variable through sweeps
}
