% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equacheck.R
\name{equa_check}
\alias{equa_check}
\title{check equalabration}
\usage{
equa_check(o02data = NULL, sweeps = NULL, dir.res = NULL,
  addon = "", comp = "chi2")
}
\arguments{
\item{o02data}{list. o02 data struct. must be provided}

\item{sweeps}{numeric. the last n sweeps that should be included in the drawing. must be provided}

\item{dir.res}{string. result directory. must be provided}

\item{addon}{string. addon string for names. default ""}

\item{comp}{string. whether plot parameter with sweep?
"chi2": only plot chi2. default
"all": plot all}
}
\value{
no return just plot
}
\description{
this function check the trend of chi^2 and parameters
this is prefered before accumulation run
it can check the chi^2~sweeps, initial_concentration~time, and kinetic parameters~time
}
\examples{

}
