% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table_reader.R
\name{exp_model_tab}
\alias{exp_model_tab}
\title{experiment specie concentration fetching(o01)}
\usage{
exp_model_tab(lines, start, end, specname_ind, power_ind, ch_ind, species)
}
\arguments{
\item{lines}{array. lines to parse}

\item{start}{array. start of the index range}

\item{end}{array. end of the index range}

\item{specname_ind}{array. the index of species part}

\item{power_ind}{array. the index of power line}

\item{ch_ind}{array. the index of the number line}

\item{species}{string. the specie to fetch}
}
\value{
array. the informaiton table
}
\description{
experiment specie concentration fetching from the table in file o01
}
\seealso{
[mod_model_tab()] for similar function on model estimation information
}
