% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formu.reac.R
\name{formu_reac}
\alias{formu_reac}
\title{formualte transcription and translation}
\usage{
formu_reac(enz, path)
}
\arguments{
\item{enz}{array. the enzyme list}

\item{path}{string. the result location}
}
\value{
no return
}
\description{
this is funciton is for producing transcription and translation reaction from a gene/enzyme list
}
