% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoge_file.R
\name{homoge_file}
\alias{homoge_file}
\title{modify output script for consistency}
\usage{
homoge_file(localpath = NULL, foldname = NULL)
}
\arguments{
\item{localpath}{string. the local working directory. must be provided}

\item{foldname}{string. the name of the project folder. must be provided}
}
\value{
print out result no return
}
\description{
makesure the output analysis script agreed with the original input file
}
