% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kine.para.refine.R
\name{kine_para_refine}
\alias{kine_para_refine}
\title{refine kinetic parameters}
\usage{
kine_para_refine(list.res.refine, range.speci, extendrag)
}
\arguments{
\item{list.res.refine}{list. the species non-specific data struct}

\item{range.speci}{list. the species specific data struct}

\item{extendrag}{numeric. extent to extend unknown range in kinetic parameters(when there is only one value in in record)}
}
\value{
list. refined parameter list
}
\description{
this function will take kinetic parameters structure and modify it
will calcualte range for all kinetic parameters and give default value for those without enough information
}
