% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_table_reader.R
\name{mod_model_tab}
\alias{mod_model_tab}
\title{model specie concentration fetching(o01)}
\usage{
mod_model_tab(lines, start, end, specname_ind, power_ind, species)
}
\arguments{
\item{lines}{array. lines to parse}

\item{start}{array. start of the index range}

\item{end}{array. end of the index range}

\item{specname_ind}{array. the index of species part}

\item{species}{string. the specie to fetch}

\item{ch_ind}{array. the index of the number line}
}
\value{
array. the informaiton table
}
\description{
model specie concentration fetching from the table in file o01
}
\seealso{
[exp_model_tab()] for similar function on model estimation information
}
