% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{numparse}
\alias{numparse}
\title{parse the parameter numbers for theta speceis/parameters}
\usage{
numparse(lines_models, this_ind, next_ind, name)
}
\arguments{
\item{lines_models}{array. lines for searching}

\item{this_ind}{array. define the range start}

\item{next_ind}{array. define the range end}

\item{name}{array. name of the theta vector}
}
\value{
list. the named parameter list
}
\description{
o02 file blocks of species parameter, reaction parameter, acceptance rate information and step size information.
}
