% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multi.spec.time.R
\name{plot_multi_spec_time}
\alias{plot_multi_spec_time}
\title{plot of time trajectory of multiple species}
\usage{
plot_multi_spec_time(o01, species.list, addon, exp, dir.res, ncol = 5,
  logtrans = FALSE)
}
\arguments{
\item{o01}{string. the location of o01}

\item{species.list}{array. the species to plot}

\item{addon}{string. addon file name}

\item{exp}{int. the experiment number to choose}

\item{dir.res}{string. result locaiton}

\item{logtrans}{bool. whether log transformation is  performed}
}
\value{
figure. the figure itself
}
\description{
plot of time trajectory of multiple species
}
