% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randloglhs.R
\name{randloglhs}
\alias{randloglhs}
\title{initial guess generate}
\usage{
randloglhs(nsample = NULL, isample = NULL, input = NULL,
  output = NULL)
}
\arguments{
\item{nsample}{int. number of replicate to produce. must be provided}

\item{isample}{int. the sample index. must be provided}

\item{input}{string. input file path. must be provided}

\item{output}{string. output file path. must be provided}
}
\value{
just change the file no return
}
\description{
this function will generate the initial guess for all parameters by lhs (Latin Hypercube sample) in log-uniform distribution
the lhs matrix will be calcualted for isample and used afterwards
}
