% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{rela_ind_screen}
\alias{rela_ind_screen}
\title{use relative locaiton between indx and indy to filter indy}
\usage{
rela_ind_screen(indx = NULL, indy = NULL, rela = ">=")
}
\arguments{
\item{indx}{array. the reference index. must be provided}

\item{indy}{array. the index to be filtered. must be provided}

\item{rela}{string. the expected relative location between indx and indy. default ">="}
}
\value{
array. the filtered indy
}
\description{
use relative locaiton between indx and indy to filter indy
}
