% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_o02}
\alias{summary_o02}
\title{summary plot for o02}
\usage{
summary_o02(o02.data = NULL, dir.res = NULL, addonname = "",
  linethick = FALSE)
}
\arguments{
\item{o02.data}{list. the struct from o02.reader. must be provided}

\item{dir.res}{string. the location for producing figures. must be provided}

\item{addonname}{string. the addon name. default ""}

\item{linethick}{bool. whether thicker lines shoulde be draw for figure. default FALSE}
}
\value{
just plot no return
}
\description{
summary plot for o02
}
