% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_reac}
\alias{summary_reac}
\title{reaction summary}
\usage{
summary_reac(path, enzpattern)
}
\arguments{
\item{path}{string. path to the reaction list file}

\item{enzpattern}{string. the pattern to search for the enzyme entity}
}
\value{
list. list containing enzyme information
}
\description{
path the reaction file location
warning on input equation file
}
