% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_exp_data.R
\name{template_exp_data}
\alias{template_exp_data}
\title{formualte the dataset (i02)}
\usage{
template_exp_data(listdata, classnum, output, dir.data, modified.file,
  zspec_wid = 0.4, sameclass = 0)
}
\arguments{
\item{listdata}{list. the list contains time and value}

\item{classnum}{int. the start class number to be considered, a new class number will be return}

\item{output}{string. ouput the add part of experiment measrument}

\item{dir.data}{string. the working/data directory}

\item{modified.file}{string. addon file name}

\item{zspec_wid}{numeric. the zspec_wid for new input data, the default is 0.4}

\item{sameclass}{int. whether different scale class for different input measurement,
the default is 0 (use different class number). 1 use the same class number}
}
\value{
int. the new class number to be used
}
\description{
formualte the dataset (i02)
}
