% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.R
\name{update_history}
\alias{update_history}
\title{record history in the history file}
\usage{
update_history(histpath = NULL, projpath = NULL)
}
\arguments{
\item{histpath}{string. the history record file. must be provided}

\item{projpath}{string. the current working project directory. must be provided}
}
\value{
no return just modify the file
}
\description{
record history in the history file
}
